refactor: consolidate ExpenseItemCard and TaxItemCard into renderItemCard function in ExpensesTab for improved code maintainability
